+++ examples/hello_world_2.py
@@ -5,7 +5,7 @@ from gladius.html5 import Html5
 from gladius.daisyui import DaisyUI
 
 g = Gladius()
-html5 = Html5(g)
+h = Html5(g)
 d = DaisyUI(g)
 
 # callbacks
@@ -19,16 +19,16 @@ async def world_button_click(button: Component, event: Event):
     world_card.content = f'World {v}'
     world_button.content = f'World {v}'
 
-page = html5.Page()
-page.head.add(title := html5.Title(content='Hello world 1'))
-page.head.add(link := html5.Link(rel='shortcut icon', type='image/png', href='/static/gladius/favicon.png'))
-page.head.add(link := html5.Link(href='https://cdn.jsdelivr.net/npm/daisyui@3.8.0/dist/full.css', rel='stylesheet', type='text/css'))
-page.head.add(script := html5.Script(src='https://cdn.tailwindcss.com'))
-page.head.add(script := html5.Script(src='https://unpkg.com/htmx.org@1.9.6'))
-page.head.add(script := html5.Script(src='https://unpkg.com/htmx.org/dist/ext/debug.js'))
-page.head.add(script := html5.Script(src='https://unpkg.com/htmx.org/dist/ext/json-enc.js'))
-page.head.add(script := html5.Script(src='https://unpkg.com/htmx.org/dist/ext/event-header.js'))
-page.head.add(script := html5.Script(src='/static/gladius/multi-path-deps.js'))
+page = h.Page()
+page.head.add(title := h.Title(content='Hello world 1'))
+page.head.add(link := h.Link(rel='shortcut icon', type='image/png', href='/static/gladius/favicon.png'))
+page.head.add(link := h.Link(href='https://cdn.jsdelivr.net/npm/daisyui@3.8.0/dist/full.css', rel='stylesheet', type='text/css'))
+page.head.add(script := h.Script(src='https://cdn.tailwindcss.com'))
+page.head.add(script := h.Script(src='https://unpkg.com/htmx.org@1.9.6'))
+page.head.add(script := h.Script(src='https://unpkg.com/htmx.org/dist/ext/debug.js'))
+page.head.add(script := h.Script(src='https://unpkg.com/htmx.org/dist/ext/json-enc.js'))
+page.head.add(script := h.Script(src='https://unpkg.com/htmx.org/dist/ext/event-header.js'))
+page.head.add(script := h.Script(src='/static/gladius/multi-path-deps.js'))
 page.body.set_attr(hx_ext='multi-path-deps').add_class('p-10')
 
 # vflex