+++ gladius/html5.py
@@ -38,6 +38,7 @@ __all__ = [
     'Source',
     'Hr',
     'Button',
+    'Main',
     'Html5',
 ]
 
@@ -73,7 +74,7 @@ class Page(Component):
         head.add(title := h.Title(content=self.title))
         head.add(link := h.Link(rel='shortcut icon', type='image/png', href=self.favicon))
         head.add(script := h.Script(src='https://unpkg.com/htmx.org@1.9.6'))
-        # head.add(script := h.Script(src='https://unpkg.com/htmx.org/dist/ext/debug.js'))
+        head.add(script := h.Script(src='https://unpkg.com/htmx.org/dist/ext/debug.js'))
         head.add(script := h.Script(src='https://unpkg.com/htmx.org/dist/ext/json-enc.js'))
         head.add(script := h.Script(src='https://unpkg.com/htmx.org/dist/ext/event-header.js'))
         head.add(script := h.Script(src='https://unpkg.com/idiomorph/dist/idiomorph-ext.min.js'))
@@ -164,6 +165,12 @@ class H3(Component):
 class H4(Component):
     tag: str = 'h4'
 
+class H5(Component):
+    tag: str = 'h5'
+
+class H6(Component):
+    tag: str = 'h6'
+
 class P(Component):
     tag: str = 'p'
 
@@ -233,6 +240,12 @@ class Hr(Component):
 class Button(Component):
     tag: str = 'button'
 
+class Main(Component):
+    tag: str = 'main'
+
+#
+# Html5 Component Library
+#
 class Html5(ComponentLibrary):
     def __init__(self, ctx: Gladius):
         super().__init__(ctx)